refactor(2368): batch C — remove deprecated public-barrel types
Drops ITechLead, IOrchestratorLegacy, IExpertFactory (orchestrate-types
flavor), createMockTechLead, OrchestrateDeps.techLead, and
OrchestrateDeps.expertFactory from the public MCP barrels. Completes
the deprecated-types side of #1986 / #2368.

This is the typed-only minor break the user authorized for the v3.0-
gated cleanup. Runtime semantics are unchanged: the cli-server-tools.ts
internal callsite now wraps the legacy Orchestrator class with
OrchestratorFactory.create('tech_lead') to produce an IOrchestrator.

ITechLead survives as an internal-only type for the SICA adapter
cascade (orchestrate-sica.ts still depends on the {execute(task)} shape
to wrap with SicaAgent). It is no longer reachable through any
src/exports/*.ts barrel — confirmed via repo-wide grep.

createMockTechLead is removed from the public surface; the mock task-
executor logic is a private helper inside createMockOrchestrator.

Migration recipes documented in the changeset.

Out of scope (deliberate, separate PRs):
- useMockTechLead config field rename → useMockOrchestrator (option-name
  break, narrower audience)
- agents/experts/task-analyzer.ts removal (separate concern)
- OrchestratorType = 'tech_lead' | ... discriminator rename (separate
  type system, not an AgentRole)

Verified:
- pnpm typecheck: clean
- pnpm lint: clean
- pnpm vitest run: 26,414 passed, 16 skipped, 1055 files

Refs #2368, #1986, #595, #759
